Refactoring
More on #22

- Get files is just like get data, where the conversion does nothing - so lets not have the same code twice.
- Pull out some utilities from the servicex file to make it "shorter"
